Moved react router from index.tsx to App.tsx
